+++ src/components/Table/ButtonContainer/DetailsButton/index.ts
@@ -1,0 +1,3 @@
+import DetailsButton from './DetailsButton';
+
+export default DetailsButton;
