+++ src/components/Table/ButtonContainer/EditButton/index.ts
@@ -1,0 +1,3 @@
+import EditButton from './EditButton';
+
+export default EditButton;
