+++ src/components/Table/ButtonContainer/index.ts
@@ -1,0 +1,3 @@
+import ButtonContainer from './ButtonContainer';
+
+export default ButtonContainer;
